# apps/quantai/openapi.yaml
openapi: 3.0.3
info:
  title: QuantAI API
  description: AI platform with agents, orchestration, memory, tools, ML pipelines, and cross-app coordination.
  version: 1.0.0
  contact:
    name: Quant Ecosystem
servers:
  - url: /api/v1
    description: QuantAI API server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request body
            statusCode:
              type: integer
              example: 400
...
security:
  - BearerAuth: []
paths:
...
  /federation/register:
    post:
      tags: [Federation]
      summary: Register a federated agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, endpoint, capabilities, trustLevel]
              properties:
                id:
                  type: string
                endpoint:
                  type: string
                  format: uri
                capabilities:
                  type: array
                  items:
                    type: string
                trustLevel:
                  type: number
                  minimum: 0
                  maximum: 1
      responses:
        '200':
          description: Agent registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /federation/call/{id}:
    post:
      tags: [Federation]
      summary: Call a federated agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                input:
                  type: string
      responses:
        '200':
          description: Agent result
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Federation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...